feat(input,widgets,render,runtime,pty): Alt+Ctrl key parsing, scroll-guarded list, scrollbar hit encoding, Presenter integration, and WS flow telemetry
Input parser (ftui-core):
- Add match arm for control characters (0x00..0x1F) after ESC prefix so
  Alt+Ctrl+A (ESC + 0x01) is correctly decoded instead of being silently
  swallowed; add inline unit tests for Alt+Ctrl+A and Alt+Backspace
- Delete standalone repro_alt_ctrl_parsing.rs (superseded by inline tests)

List widget (ftui-widgets):
- Add scroll_into_view_requested flag to ListState; auto-scroll into
  viewport only fires when flag is set (on select/next/prev), cleared
  after render applies it — fixes bug where mouse wheel scroll was
  immediately overridden by forced viewport centering
- Add regress_list_scroll.rs regression test

Scrollbar widget (ftui-widgets):
- Change hit data encoding from [8:part][56:track_pos] to
  [8:part][28:track_len][28:track_pos] so mouse handlers can do
  accurate proportional jumps without using viewport_length as proxy
- Account for begin/end button symbols as separate regions outside
  effective track when computing thumb position
- Guard against track_len == 0 in click and drag handlers
- Add repro_table_header_width.rs for FitContent column header bug

Presenter (ftui-render):
- Add viewport_offset_y field for inline mode row offset
- Add set_viewport_offset_y(), writer_mut(), prepare_runs() public API
- Rename emit_runs_reuse() to emit_diff_runs() (pub), remove dead emit_runs()

TerminalWriter (ftui-runtime):
- Replace internal Option<CountingWriter<BufWriter<W>>> with Presenter<W>
  for diff emission through Presenter's optimized path
- render_frame() now uses presenter.reset/set_viewport_offset_y/
  prepare_runs/emit_diff_runs pipeline
- gc() accepts optional extra_buffer to avoid premature grapheme pool
  collection when render buffer is pending

WebSocket bridge (ftui-pty):
- Add emit_flow_control_event() and emit_flow_control_stall_if_transitioned()
  helpers for structured telemetry (direction, action, window/queued bytes)
- Emit replenish events on input flow control path
- Add 2 tests for payload fields and rising-edge stall transition

Golden frame hash tests (ftui-harness):
- BorderType::Thick renamed to BorderType::Heavy
- Grid tests updated from flat cells iteration to explicit layout.cell(r,c)
- Replace format!() allocations with static string slices in hot paths

Other fixes:
- ftui-tty: change poll timeout from u16 to i32 (u16 silently truncates >65s)
- ftui-layout: suppress unused variable warning (_i)
- Update beads issue tracker metadata

+++ crates/ftui-tty/src/lib.rs
@@ -423,7 +423,9 @@ impl TtyEventSource {
                 tty.as_fd(),
                 nix::poll::PollFlags::POLLIN,
             )];
-            let timeout_ms: u16 = timeout.as_millis().try_into().unwrap_or(u16::MAX);
+            // Use i32 for timeout to allow values > 65s (up to ~24 days).
+            // poll(2) takes milliseconds as a signed int.
+            let timeout_ms: i32 = timeout.as_millis().try_into().unwrap_or(i32::MAX);
             match nix::poll::poll(&mut poll_fds, nix::poll::PollTimeout::from(timeout_ms)) {
                 Ok(n) => n,
                 Err(nix::errno::Errno::EINTR) => return Ok(false),
